feat: added second and millisecond support for time picker
