+++ src/lib/DatePicker.svelte
@@ -1,5 +1,6 @@
 <script lang="ts">
   import { browser } from '$app/environment'
+  import TimePicker from './TimePicker.svelte'
   import { getMonthLength, getCalendarDays, type CalendarDay } from './date-utils.js'
   import { getInnerLocale, type Locale } from './locale.js'
   import { createEventDispatcher, onMount } from 'svelte'
@@ -211,117 +212,6 @@
    * Time picker
    */
   export let timePrecision: null | 'minute' | 'second' | 'millisecond'
-  $: timePickerState = ''
-  let timePickerActiveField: null | HTMLInputElement
-  $: timePickerActiveField = null
-
-  let hourInput: HTMLInputElement
-  let minuteInput: HTMLInputElement
-
-  function setTimePickerActiveField(e: Event) {
-    if (e.target instanceof HTMLInputElement) {
-      timePickerActiveField = e.target
-    }
-  }
-
-  function isDigitKey(key: string) {
-    var digitPattern = /^[0-9]$/
-    return digitPattern.test(key)
-  }
-
-  const withinMinuteRange = (m: number) => m >= 0 && m <= 59
-  const withinHourRange = (h: number) => h >= 0 && h <= 23
-
-  function setHours(newHour: number | typeof NaN) {
-    if (withinHourRange(newHour)) {
-      hourInput.value = `${newHour < 10 ? '0' : ''}${newHour}`
-      browseDate.setHours(newHour)
-      browse(browseDate)
-    }
-  }
-
-  function setMinutes(newMinute: number) {
-    if (withinMinuteRange(newMinute)) {
-      minuteInput.value = `${newMinute < 10 ? '0' : ''}${newMinute}`
-
-      browseDate.setMinutes(newMinute)
-      browse(browseDate)
-    }
-  }
-
-  function setInitilState(el: typeof timePickerActiveField) {
-    if (el) {
-      timePickerState = el.value == '00' ? '' : el.value
-    }
-  }
-  function setTime(value: string) {
-    const valueInt = value == '' ? 0 : parseInt(value, 10)
-    if (hourInput && timePickerActiveField == hourInput) {
-      setHours(valueInt)
-    } else if (minuteInput && timePickerActiveField == minuteInput) {
-      setMinutes(valueInt)
-    }
-  }
-  $: {
-    setInitilState(timePickerActiveField)
-  }
-  $: {
-    if (browser) {
-      setTime(timePickerState)
-    }
-  }
-  function setTimePickerState(key: string) {
-    const digit = parseInt(key, 10)
-
-    if (timePickerActiveField == hourInput && withinHourRange(digit)) {
-      timePickerState = key
-    } else if (timePickerActiveField == minuteInput && withinMinuteRange(digit)) {
-      timePickerState = key
-    }
-  }
-  function increaseStrNumber(str: string) {
-    if (str === '') {
-      return '1'
-    } else {
-      const number = parseInt(str, 10)
-      const incrementedNumber = number + 1
-      return String(incrementedNumber)
-    }
-  }
-
-  function decreaseStrNumber(str: string) {
-    if (str === '' || str === '0') {
-      return '0'
-    } else {
-      const number = parseInt(str, 10)
-      const decrementedNumber = number - 1
-      return String(decrementedNumber)
-    }
-  }
-
-  function timePickerKeydown(e: KeyboardEvent) {
-    e.preventDefault()
-    if (timePickerActiveField == hourInput) {
-      const keysToMoveNext = Array.from(';:-,.').concat(['ArrowRight', 'Tab'])
-      if (keysToMoveNext.includes(e.key) && !(e.shiftKey && e.key == 'Tab')) {
-        minuteInput.focus()
-      }
-    } else if ((e.key == 'Tab' && e.shiftKey) || e.key == 'ArrowLeft') {
-      hourInput.focus()
-    }
-    if (isDigitKey(e.key)) {
-      setTimePickerState(timePickerState + e.key)
-    } else if (e.key == 'ArrowUp') {
-      console.log('increase', increaseStrNumber(timePickerState))
-      setTimePickerState(increaseStrNumber(timePickerState))
-    } else if (e.key == 'ArrowDown') {
-      setTimePickerState(decreaseStrNumber(timePickerState))
-    }
-
-    if (e.key == 'Backspace' && timePickerState != '') {
-      timePickerState = timePickerState.slice(0, -1)
-    }
-  }
 </script>
 
 <!-- svelte-ignore a11y-no-noninteractive-tabindex -->
@@ -436,31 +326,11 @@
         {/each}
       </div>
     {/each}
-    {#if timePrecision != null}
-      <div class="timepicker" on:keydown={timePickerKeydown}>
-        <input
-          bind:this={hourInput}
-          on:focus={setTimePickerActiveField}
-          aria-label="hours (24hr clock)"
-          type="number"
-          class="timepicker-input timepicker-hour"
-          id="hour24-input-0"
-          min="0"
-          max="23"
-          value="00"
-        /><span class="timepicker-divider-text">:</span><input
-          on:focus={setTimePickerActiveField}
-          aria-label="minutes"
-          bind:this={minuteInput}
-          value="00"
-          type="number"
-          min="0"
-          max="59"
-          class="timepicker-input timepicker-minute"
-          id="minute-input-0"
-        />
-      </div>
-    {/if}
+
+    <TimePicker timePrecision={null} bind:browseDate {browse} />
+    <TimePicker timePrecision={'minute'} bind:browseDate {browse} />
+    <TimePicker timePrecision={'second'} bind:browseDate {browse} />
+    <TimePicker timePrecision={'millisecond'} bind:browseDate {browse} />
   </div>
 </div>
 
@@ -597,34 +467,5 @@
       color: var(--date-picker-selected-color, inherit)
       background: var(--date-picker-selected-background, rgba(2, 105, 247, 0.2))
       border: 2px solid var(--date-picker-highlight-border, #0269f7)
-  .timepicker
-    display: flex
-    justify-content: center
-    align-items: center
-    margin-block:0.5rem
-    padding: 1px
-    width: fit-content
-    margin-inline: auto
-    border: 1px solid rgba(108, 120, 147, 0.3)
-    border-radius: 5px
-    .timepicker-divider-text
-      padding-inline:2px
-    .timepicker-input
-      max-width:1.5rem
-      width: fit-content
-      border: none
-      outline: none
-      height: auto
-      -moz-appearance: textfield
-      font-family: inherit
-      text-align: center
-      caret-color: transparent
-      &:focus-visible
-        outline: none
-        background-color: rgba(#808080, 0.08)
-      &::input::-webkit-outer-spin-button, &::input::-webkit-outer-spin-button
-        -webkit-appearance: none
-        margin: 0
-
 
 </style>
